Set frontend output to standalone and update Dockerfile
In the 'next.config.js' file, the output has been set to 'standalone'. This change affects the Dockerfile, which now includes steps to create the static files in the correct directories for standalone use. Additionally, the production environment has been set and the command to start the server has changed to run off 'server.js'.