Add FRONTEND_HOSTNAME to config.js and use it in index.js
This diff adds a new constant in the configuration file, FRONTEND_HOSTNAME, and uses it in index.js. Previously, the hostname for the frontend was hard-coded. This change allows the front-end hostname to be set dynamically from the environment, making the system more configurable and easier to adapt to different deployment environments.