Update README.md with project description and contact badges
Removed placeholder text and replaced with a deeper description of the project, providing more detailed insight into what the platform does and why it is valuable for software development teams.
The placeholders for LinkedIn and Discord contact methods were replaced by actual badges which directly link to the respective platforms. This offers a clean and direct way for users to reach out or connect.
Fixed a typo in the docker cli link.