Refine instructions in README.md for reverse proxy setup
The wording in the section about reverse proxy setup has been clarified. This change is to help users avoid misunderstandings when adding configuration to their proxy host. The term "headers" has been replaced to reflect accurately that it's not only headers that need to be added, preventing possible `5XX` errors.