readme (#2)
* Update README.md with more details on environment variables

This commit removes redundant line gaps from the introductory paragraph and expands the "Parameters" section in README.md. The added information provides directions on using backend and frontend environment variables. This is crucial for users to configure the application to suit their needs and facilitate its usage.

* Update Readme with new environment variables

Added new environment variables to the README.md to accommodate recently introduced features related to authentication and keycloak. This step is crucial for better user understanding and environment setup. The added variables are related to the login provider, nextauth url, secrets, and keycloak authentication details.

* Update README's environment variables requirements

The README.md file was edited to clearly identify 'AUTH_KEYCLOAK_ID', 'AUTH_KEYCLOAK_SECRET', and 'AUTH_KEYCLOAK_ISSUER' as mandatory fields when 'NEXT_PUBLIC_DEFAULT_LOGIN_PROVIDER' is set to 'keycloak.' Previously, these requirements were not very clear, leading to potential configuration errors.

* Add Backend and Frontend Ports to README

Added sections describing the Backend and Frontend Ports to README.md. The ports detail which Ports are being used by the API, Websocket server, and UI. This clarifies operation and facilitates better understanding of the system by new users or developers.

* Add License and Release badges to README

Two new badges were added to the README file: License and Release. These badges provide quick and easy access to important repository information, improving the user experience and clarity of the project status.

* Update README with additional environment variables

An expanded list of environment variables for both backend and frontend have been added to the README. This update aims at enhancing communication regarding both the backend's HTTP and asset transfer and the frontend's provider login information, providing comprehensive information to facilitate setup and integration.

* Add reverse proxy instructions to README

Added a new section in the README.md on how to correctly set up a reverse proxy such as nginx. This update is necessary as nextauth payloads can be large when redirecting back from the oauth provider and without these settings, the users might encounter unexpected `5XX` errors.

* "Replace markdown image syntax with HTML in README"

Replaced the markdown syntax for including images with HTML to gain control over image size in README.md. Necessary for displaying the images in an appropriate dimension and improve readability of the document.

* Update image size in README.md

The size of the `npm-advanced-tab` image has been updated in the README.md file to increase visibility. The width has been changed from 200 pixels to 250 pixels for better readability and user experience.

* Refine instructions in README.md for reverse proxy setup

The wording in the section about reverse proxy setup has been clarified. This change is to help users avoid misunderstandings when adding configuration to their proxy host. The term "headers" has been replaced to reflect accurately that it's not only headers that need to be added, preventing possible `5XX` errors.

* Add Minio section to README

A new section describing the use of Minio, a self-hosted storage solution was added to the README. This addition provides important information for users about generation of AWS keys, use of these keys, and underscores the need for adherence to security best practices.

* Update README with production setup for docker-compose

The README file has been significantly updated to include a comprehensive setup for docker-compose in a production environment. This setup guide includes detailed configuration instructions for vital services such as PostgreSQL, Minio, and the openpro backend and frontend. The changes were made to better guide users through setting up a production-ready environment.