"Added Python and pip to Docker backend"
This commit introduces Python and its package manager pip to the Docker backend as part of the build. The PYTHONUNBUFFERED environment variable is unbuffered binary stdout and stderr to help see what's going on in real-time during building and deployment. It also installs g++ and makes necessary for Python packages requiring binary compilation.