Update API URL and CSS properties in Editor
API URL has been updated to use the service config in order to remove the hard coded localhost URL within Editor component. Also, CSS properties in the section element were rearranged for better readability. This makes the code more scalable and easier to manage in a case of potential changes to the API or server location.