Update websocket protocol in Editor component
The url used for initiating a websocket connection in the Editor component has been updated to dynamically use 'wss' or 'ws' based on the NEXT_PUBLIC_API_URL protocol. This change ensures secure websocket connections ('wss') are used when the API URL is using HTTPS, improving the security of the data communication.