Implement realtime collaborative editing with Yjs
Realtime collaborative editing has been implemented using Yjs. This involved the updating of the component structure in the frontend to use the HocuspocusProvider and changes in the backend with @hocuspocus/extension-database integration for fetching and storing document state. Additionally, a 'description_raw' column has been added to the issues table in the database to manage collaborative inputs.