Update handling of comments in Issue Modal and backend
The update includes a change in the documentName prop for the Editor component in Issue Modal to make it more dynamic. The backend is updated to handle 'comment' entities in database queries. Also, a new 'commentRaw' field has been added to the 'issue-comment' model in the database schema to store comment content.