Add issue linking and retrieval functionality
The code now allows Issues to be linked with each other and also fetches linked issues as part of the issue information. Changes have been made on the back-end database models to accommodate this functionality, while the GraphQL implementation has been updated to carry the linked issue details alongside the main issue information.