Remove default empty string for token
The code for handling the absence of an authorization header has been changed. The previous implementation used an empty string as a default value for the token if it was not found in the request headers. Now, the 'token' variable will be set to 'undefined' if the 'authorization' header is not present.