Add linked issues component to issue modal
A new component called 'IssueLinkedIssues' has been added to the issue modal, which displays the list of issues linked to the current one. It includes sample data which needs to be replaced by live data from the backend in the future. Additionally, search and searchOperator fields have been added to the QueryIssueInput type.