Add issue link creation feature to IssueModal
The changes introduce a new feature allowing users to input a link connection between issues directly in the IssueModal component. Two dropdowns were added: one for defining the link type, and another for selecting the issue to be linked. The buttons for creating the link, cancelling the operation, or searching for an issue were also incorporated.