Add vectorSearch field to Issue model
The field 'vectorSearch' of the type TSVECTOR has been added to the Issue model in the database. The search functionality in resolvers.js has also been updated to utilize this new field for more efficient and accurate search results. The commit removes the redundant condition which checks if either projectId or id is not provided.