Add project details to issues and implement issue search
This update includes the addition of 'project' details to the 'Issue' GraphQL type, enabling these details to be accessed with the issue. Additionally, the "LinkIssueSearch" function in 'IssueLinkedIssues.tsx' has been enhanced to actually perform a query and return matching issues. To improve performance, a debounce function has been implemented to delay the search trigger by 300ms.