Add functionality to open linked issues
A new feature is implemented to enable opening of linked issues from the issue modal list. A function named 'handleOpenIssue' is added which takes an issue id and navigates to the respective issue detail page. Click actions are also added to the issue id and title list items to trigger this function.