Add Issue Link Creation Feature in IssueModal Component (#15)
* Add issue link creation feature to IssueModal

The changes introduce a new feature allowing users to input a link connection between issues directly in the IssueModal component. Two dropdowns were added: one for defining the link type, and another for selecting the issue to be linked. The buttons for creating the link, cancelling the operation, or searching for an issue were also incorporated.

* Add vector search to issues in database

The main modification involves adding a vector_search column and corresponding index to the 'issues' table in sequelize migrations. This newly added TSVECTOR type column is used for improved full-text search. Meanwhile, the resolvers.js file has been updated to use the exact match operator for the 'id' search field and case-insensitive like for the 'title' field.

* Add vectorSearch field to Issue model

The field 'vectorSearch' of the type TSVECTOR has been added to the Issue model in the database. The search functionality in resolvers.js has also been updated to utilize this new field for more efficient and accurate search results. The commit removes the redundant condition which checks if either projectId or id is not provided.

* Refactor migration script for issue vector search

The code has been simplified by switching to async/await pattern from promises for better readability and ease of understanding. This was done in the `up` and `down` functions within the `add-issue-vector-search` migrations script. Unnecessary console logs have also been removed.

* Add project details to issues and implement issue search

This update includes the addition of 'project' details to the 'Issue' GraphQL type, enabling these details to be accessed with the issue. Additionally, the "LinkIssueSearch" function in 'IssueLinkedIssues.tsx' has been enhanced to actually perform a query and return matching issues. To improve performance, a debounce function has been implemented to delay the search trigger by 300ms.

* Add comment about CSS bug in IssueLinkedIssues component

A comment has been added to the IssueLinkedIssues file in the frontend components to highlight an existing CSS bug. The bug is causing a problem with how the dropdown appears in a modal. This comment will serve as a reminder to fix the issue in the future.

* Optimize key prop in IssueLinkedIssues component

The update simplifies the mapped Listbox.Option elements in the IssueLinkedIssues component by directly using 'link.value' as the key prop instead of 'linkIdx'. This can potentially optimize react re-renderings for this component.

* Add functionality to create and delete issue links

This commit introduces methods for creating and deleting issue links. These new functionalities are introduced in the backend and connected to the frontend via GraphQL queries and mutations. Changes also include type definitions and model adjustments for proper link handling.

* Add functionality to delete issue links

A delete issue link mutation has been implemented on both the frontend and backend. On the frontend, it's represented as a 'X' icon in the issue links section. On the backend, it's handled by destroying the corresponding record in the IssueLinks model in the database.

* Add issue link creation and deletion handlers

A new type named 'IssueLink' has been defined to accommodate link data including 'issueId', 'linkedIssueId', and 'linkType'. Two new functions have been added: 'handleDeleteIssueLink' and 'handleCreateIssueLink' for issue link handling on deletion and creation events respectively. In these functions, mutation calls for creating and deleting issue links have been abstracted.

* Add functionality to open linked issues

A new feature is implemented to enable opening of linked issues from the issue modal list. A function named 'handleOpenIssue' is added which takes an issue id and navigates to the respective issue detail page. Click actions are also added to the issue id and title list items to trigger this function.

* Refactor IssueLinkedIssues component into multiple files

The codebase has been refactored to split the complex IssueLinkedIssues component into separate, more manageable elements. This change improves readability and maintainability, as each subcomponent can now be understood and manipulated independently. The subcomponents were moved into their respective files: LinkIssueTypeDropdown, LinkIssueSearch, and the Index file for the IssueLinkedIssues.

* Refactor code in LinkIssueSearch component

The code in the LinkIssueSearch component has been refactored for improved readability and efficiency. The sequence of useState and useLazyQuery hooks have been reordered, creating a logical, easy-to-follow flow within the component. The code also optimizes performance by implementing a debounce function on handleChange for efficient search results.

* Remove unused setter in issue-links model

The unused setter function for the "linkTypeInverted" field in the issue-links model has been removed. This cleanup further simplifies the model and enhances code readability.

* Add DeleteIssueLink functionality in GraphQL

The code now includes the DeleteIssueLink mutation in GraphQL schema. The new feature enables us to remove the issue links in the application. The respective GraphQL documents have been generated and added to the schema.