Update UI color scheme in IssueModal
Altered several CSS class definitions in the LinkIssueSearch component of the IssueModal. These changes focus on updating the UI color scheme, transitioning from a white/grey and teal scheme to a "surface-overlay" and "primary" color scheme. The outcome maintains the accessibility in the color contrasts, while creating a more unified visual user experience.