Add project settings page and user management feature
This commit introduces a project settings page in the frontend. In addition, it also adds the ability to add or remove users from a project in the backend. This includes accompanying GraphQL query and mutation setup. Now, users can be associated with a project and be managed via the settings page.