Implementing Mention Feature in Editor (#19)
* Add mention feature to editor

This commit introduces a mention feature to the editor. The new feature is implemented using the @tiptap/mention and @tiptap/suggestion extensions. Additionally, a MentionList component has been added to present a selectable list of user names when '@' is typed in the editor.

* Change key prop in MentionList from index to item.id

An amendment was done in the frontend/components/Editor/extensions/MentionList.tsx. Instead of using the index as a key prop for the items, item.id is now being utilized. This change could provide a performance improvement as using the index as a key could lead to issues if the items list changes dynamically.