Add user settings and refactor user formatting
This commit adds a new 'settings' field to user related GraphQL types, queries and mutations on both frontend and backend, while maintaining the same response shape by using a helper function. In addition, 'js-confetti' and 'react-canvas-confetti' were added to the frontend package.json and a new migrations file was created to update the user table in the database. Lastly, a small visual style adjustment was made on the avatar upload component.