Optimize formatting and data loading in resolvers (#28)
The primary action taken in this commit was the optimization of GraphQL resolvers. This was done by including a the 'formatUserForGraphql' function in issues, projects, and issue comments resolvers which helped streamline the process of formatting users for GraphQL. In addition, the DataLoader was primed with 'findIssueComment', improving data loading efficiency. Lastly, code related to GraphQL after file write hooks and mappers were revised in 'codegen.ts'.