Refactor code and enhance board link generation
The existing code in RecentProjectsCards.tsx has been refactored for more efficient project card data formatting. This was achieved by restructuring and resorting the imports and destructuring the project object. A significant enhancement has been made to the way in which the board link is generated, using `boardId` instead of the hard-coded value of '1'. An option for additional querying options has also been added in the board resolver.